pyproject.toml
[tool.poetry]
name = "ch-card-id-api"
version = "0.1.0"
description = ""
authors = ["Arsenii Kuznetsov <sdlmerr@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
torch = "=1.1.0"
torchvision = "=0.2.2"
flask = "^1.1"
gunicorn = "^19.9"
numpy = "=1.16.3"
pillow = "=6.0.0"
attrs = "=19.1.0"
pyyaml = "=5.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
black = "=19.3b0"
isort = "^4.3"
flake8 = "^3.7"
pre-commit = "^1.18"
mypy = "^0.720.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120
target-version = ['py36']
